+++ rust/cubesqlplanner/cubesqlplanner/src/planner/planners/multi_stage/member_query_planner.rs
@@ -12,12 +12,14 @@ use crate::planner::planners::{
 use crate::planner::query_tools::QueryTools;
 use crate::planner::sql_evaluator::sql_nodes::SqlNodesFactory;
 use crate::planner::sql_evaluator::ReferencesBuilder;
+use crate::planner::sql_templates::PlanSqlTemplates;
 use crate::planner::QueryProperties;
 use crate::planner::{BaseDimension, BaseMeasure, BaseMember, BaseMemberHelper, BaseTimeDimension};
 use cubenativeutils::CubeError;
 use itertools::Itertools;
 use std::collections::HashMap;
 use std::rc::Rc;
+
 pub struct MultiStageMemberQueryPlanner {
     query_tools: Rc<QueryTools>,
     _query_properties: Rc<QueryProperties>,
@@ -265,24 +267,28 @@ impl MultiStageMemberQueryPlanner {
             Some(root_alias.clone()),
         );
         for (i, input) in inputs.iter().enumerate().skip(1) {
-            let left_alias = format!("q_{}", i - 1);
             let right_alias = format!("q_{}", i);
             let left_schema = cte_schemas.get(&inputs[i - 1]).unwrap().clone();
             let cte_schema = cte_schemas.get(input).unwrap().clone();
             let conditions = dimensions
                 .iter()
                 .map(|dim| {
-                    let alias_in_left_query = left_schema.resolve_member_alias(dim);
-                    let left_ref = Expr::Reference(QualifiedColumnName::new(
-                        Some(left_alias.clone()),
-                        alias_in_left_query,
-                    ));
-                    let alias_in_right_query = cte_schema.resolve_member_alias(dim);
-                    let right_ref = Expr::Reference(QualifiedColumnName::new(
-                        Some(right_alias.clone()),
-                        alias_in_right_query,
-                    ));
-                    (left_ref, right_ref)
+                    (0..i)
+                        .map(|left_alias| {
+                            let left_alias = format!("q_{}", left_alias);
+                            let alias_in_left_query = left_schema.resolve_member_alias(dim);
+                            let left_ref = Expr::Reference(QualifiedColumnName::new(
+                                Some(left_alias.clone()),
+                                alias_in_left_query,
+                            ));
+                            let alias_in_right_query = cte_schema.resolve_member_alias(dim);
+                            let right_ref = Expr::Reference(QualifiedColumnName::new(
+                                Some(right_alias.clone()),
+                                alias_in_right_query,
+                            ));
+                            (left_ref, right_ref)
+                        })
+                        .collect()
                 })
                 .collect_vec();
             let on = JoinCondition::new_dimension_join(conditions, true);
@@ -381,6 +387,7 @@ impl MultiStageMemberQueryPlanner {
             let full_key_aggregate_planner = FullKeyAggregateQueryPlanner::new(
                 cte_query_properties.clone(),
                 node_factory.clone(),
+                PlanSqlTemplates::new(self.query_tools.templates_render()),
             );
             let subqueries = multiplied_measures_query_planner.plan_queries()?;
             let result = full_key_aggregate_planner.plan(subqueries, vec![])?;