+++ rust/cubesqlplanner/cubesqlplanner/src/planner/query_properties.rs
@@ -2,6 +2,7 @@ use super::filter::compiler::FilterCompiler;
 use super::query_tools::QueryTools;
 use super::{BaseDimension, BaseMeasure, BaseMember, BaseMemberHelper, BaseTimeDimension};
 use crate::cube_bridge::base_query_options::BaseQueryOptions;
+use crate::cube_bridge::join_definition::JoinDefinition;
 use crate::plan::{Expr, Filter, FilterItem, MemberExpression};
 use crate::planner::sql_evaluator::collectors::{
     collect_multiplied_measures, has_cumulative_members, has_multi_stage_members,
@@ -31,7 +32,7 @@ impl OrderByItem {
     }
 }
 
-#[derive(Default, Clone)]
+#[derive(Default, Clone, Debug)]
 pub struct FullKeyAggregateMeasures {
     pub multiplied_measures: Vec<Rc<BaseMeasure>>,
     pub regular_measures: Vec<Rc<BaseMeasure>>,
@@ -62,6 +63,7 @@ pub struct QueryProperties {
     query_tools: Rc<QueryTools>,
     ignore_cumulative: bool,
     ungrouped: bool,
+    multi_fact_join_groups: Vec<(Rc<dyn JoinDefinition>, Vec<Rc<BaseMeasure>>)>,
 }
 
 impl QueryProperties {
@@ -127,9 +129,6 @@ impl QueryProperties {
         let (dimensions_filters, time_dimensions_filters, measures_filters) =
             filter_compiler.extract_result();
 
-        let all_join_hints = evaluator_compiler.join_hints()?;
-        let join = query_tools.join_graph().build_join(all_join_hints)?;
-        query_tools.cached_data_mut().set_join(join);
         //FIXME may be this filter should be applied on other place
         let time_dimensions = time_dimensions
             .into_iter()
@@ -157,6 +156,16 @@ impl QueryProperties {
         };
         let ungrouped = options.static_data().ungrouped.unwrap_or(false);
 
+        let multi_fact_join_groups = Self::compute_join_multi_fact_groups(
+            query_tools.clone(),
+            &measures,
+            &dimensions,
+            &time_dimensions,
+            &time_dimensions_filters,
+            &dimensions_filters,
+            &measures_filters,
+        )?;
+
         Ok(Rc::new(Self {
             measures,
             dimensions,
@@ -170,6 +179,7 @@ impl QueryProperties {
             query_tools,
             ignore_cumulative: false,
             ungrouped,
+            multi_fact_join_groups,
         }))
     }
 
@@ -193,6 +203,16 @@ impl QueryProperties {
             order_by
         };
 
+        let multi_fact_join_groups = Self::compute_join_multi_fact_groups(
+            query_tools.clone(),
+            &measures,
+            &dimensions,
+            &time_dimensions,
+            &time_dimensions_filters,
+            &dimensions_filters,
+            &measures_filters,
+        )?;
+
         Ok(Rc::new(Self {
             measures,
             dimensions,
@@ -206,7 +226,119 @@ impl QueryProperties {
             query_tools,
             ignore_cumulative,
             ungrouped,
+            multi_fact_join_groups,
         }))
+    }
+
+    pub fn compute_join_multi_fact_groups_with_measures(
+        &self,
+        measures: &Vec<Rc<BaseMeasure>>,
+    ) -> Result<Vec<(Rc<dyn JoinDefinition>, Vec<Rc<BaseMeasure>>)>, CubeError> {
+        Self::compute_join_multi_fact_groups(
+            self.query_tools.clone(),
+            measures,
+            &self.dimensions,
+            &self.time_dimensions,
+            &self.time_dimensions_filters,
+            &self.dimensions_filters,
+            &self.measures_filters,
+        )
+    }
+
+    pub fn compute_join_multi_fact_groups(
+        query_tools: Rc<QueryTools>,
+        measures: &Vec<Rc<BaseMeasure>>,
+        dimensions: &Vec<Rc<BaseDimension>>,
+        time_dimensions: &Vec<Rc<BaseTimeDimension>>,
+        time_dimensions_filters: &Vec<FilterItem>,
+        dimensions_filters: &Vec<FilterItem>,
+        measures_filters: &Vec<FilterItem>,
+    ) -> Result<Vec<(Rc<dyn JoinDefinition>, Vec<Rc<BaseMeasure>>)>, CubeError> {
+        let dimensions_join_hints = query_tools
+            .cached_data_mut()
+            .join_hints_for_base_member_vec(&dimensions)?;
+        let time_dimensions_join_hints = query_tools
+            .cached_data_mut()
+            .join_hints_for_base_member_vec(&time_dimensions)?;
+        let time_dimensions_filters_join_hints = query_tools
+            .cached_data_mut()
+            .join_hints_for_filter_item_vec(&time_dimensions_filters)?;
+        let dimensions_filters_join_hints = query_tools
+            .cached_data_mut()
+            .join_hints_for_filter_item_vec(&dimensions_filters)?;
+        let measures_filters_join_hints = query_tools
+            .cached_data_mut()
+            .join_hints_for_filter_item_vec(&measures_filters)?;
+
+        let mut dimension_and_filter_join_hints_concat = Vec::new();
+
+        dimension_and_filter_join_hints_concat.extend(dimensions_join_hints.into_iter());
+        dimension_and_filter_join_hints_concat.extend(time_dimensions_join_hints.into_iter());
+        dimension_and_filter_join_hints_concat
+            .extend(time_dimensions_filters_join_hints.into_iter());
+        dimension_and_filter_join_hints_concat.extend(dimensions_filters_join_hints.into_iter());
+        // TODO This is not quite correct. Decide on how to handle it. Keeping it here just to blow up on unsupported case
+        dimension_and_filter_join_hints_concat.extend(measures_filters_join_hints.into_iter());
+
+        let measures_to_join = if measures.is_empty() {
+            let join = query_tools
+                .cached_data_mut()
+                .join_by_hints(dimension_and_filter_join_hints_concat.clone(), |hints| {
+                    query_tools.join_graph().build_join(hints)
+                })?;
+            vec![(Vec::new(), join)]
+        } else {
+            measures
+                .iter()
+                .map(|m| -> Result<_, CubeError> {
+                    let measure_join_hints = query_tools
+                        .cached_data_mut()
+                        .join_hints_for_member(m.member_evaluator())?;
+                    let join = query_tools.cached_data_mut().join_by_hints(
+                        vec![measure_join_hints]
+                            .into_iter()
+                            .chain(dimension_and_filter_join_hints_concat.clone().into_iter())
+                            .collect::<Vec<_>>(),
+                        |hints| query_tools.join_graph().build_join(hints),
+                    )?;
+                    Ok((vec![m.clone()], join))
+                })
+                .collect::<Result<Vec<_>, _>>()?
+        };
+        Ok(measures_to_join
+            .into_iter()
+            .into_group_map_by(|(_, (key, _))| key.clone())
+            .into_values()
+            .map(|measures_and_join| {
+                (
+                    measures_and_join.iter().next().unwrap().1 .1.clone(),
+                    measures_and_join
+                        .into_iter()
+                        .flat_map(|m| m.0)
+                        .collect::<Vec<_>>(),
+                )
+            })
+            .collect())
+    }
+
+    pub fn is_multi_fact_join(&self) -> bool {
+        self.multi_fact_join_groups.len() > 1
+    }
+
+    pub fn simple_query_join(&self) -> Result<Rc<dyn JoinDefinition>, CubeError> {
+        if self.multi_fact_join_groups.len() != 1 {
+            return Err(CubeError::internal(format!(
+                "Expected just one multi-fact join group for simple query but got multiple: {}",
+                self.multi_fact_join_groups
+                    .iter()
+                    .map(|(_, measures)| format!(
+                        "({})",
+                        measures.iter().map(|m| m.full_name()).join(", ")
+                    ))
+                    .join(", ")
+            )));
+        }
+        Ok(self.multi_fact_join_groups.iter().next().unwrap().0.clone())
     }
 
     pub fn measures(&self) -> &Vec<Rc<BaseMeasure>> {
@@ -390,6 +522,7 @@ impl QueryProperties {
         let full_aggregate_measure = self.full_key_aggregate_measures()?;
         if full_aggregate_measure.multiplied_measures.is_empty()
             && full_aggregate_measure.multi_stage_measures.is_empty()
+            && !self.is_multi_fact_join()
         {
             Ok(true)
         } else {
@@ -413,9 +546,18 @@ impl QueryProperties {
             if has_multi_stage_members(m.member_evaluator(), self.ignore_cumulative)? {
                 result.multi_stage_measures.push(m.clone())
             } else {
-                for item in
-                    collect_multiplied_measures(self.query_tools.clone(), m.member_evaluator())?
-                {
+                let join = self
+                    .compute_join_multi_fact_groups_with_measures(&vec![m.clone()])?
+                    .iter()
+                    .next()
+                    .expect("No join groups returned for single measure multi-fact join group")
+                    .0
+                    .clone();
+                for item in collect_multiplied_measures(
+                    self.query_tools.clone(),
+                    m.member_evaluator(),
+                    join,
+                )? {
                     if item.multiplied {
                         result.multiplied_measures.push(item.measure.clone());
                     } else {